Fix border for list items
