+++ tests/fixtures/paragraphs_with_borders.html
@@ -61,3 +61,25 @@
         </td>
     </tr>
 </table>
+<div style="border:3.0pt solid #FFFFFF;padding:1pt 4pt 1pt 4pt"><p>White border</p></div>
+<ol class="pydocx-list-style-type-decimal">
+    <li>
+        <div style="border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">XXX</div>
+        <ol class="pydocx-list-style-type-lowerLetter">
+            <li>
+                <div style="border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">XXX1</div>
+            </li>
+            <li>
+                <div style="border:3.0pt solid #000000;border-top:0;padding:1pt 4pt 1pt 4pt">XXX2</div>
+            </li>
+        </ol>
+    </li>
+    <li>
+        <div style="border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
+            <span style="border:3.0pt solid #70AD47">YYY</span>
+        </div>
+    </li>
+    <li>
+        <span style="border:3.0pt solid #FF0000">ZZZ</span>
+    </li>
+</ol>
