+++ tests/fixtures/paragraphs_with_borders_and_shading.html
@@ -1,67 +1,66 @@
-<div style="background-color:#FF0000">
-    <p>AAA</p>
-    <p>BBB</p>
-</div>
-<div style="background-color:#00B050">
-    <p>CCC</p>
-    <p>DDD</p>
-</div>
+<div style="background-color:#FF0000"><p>AAA</p>
+    <p>BBB</p></div>
+<div style="background-color:#00B050"><p>CCC</p>
+    <p>DDD</p></div>
 <p>EEE <span style="background-color:#7030A0">FFF</span> <span
-        style="background-color:#7030A0">GGG</span>
-</p>
-<div style="background-color:#ED7D31;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
-    <p>HHH</p>
-    <p>III</p>
-</div>
+        style="background-color:#7030A0">GGG</span></p>
+<div style="background-color:#ED7D31;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt"><p>
+    HHH</p>
+    <p>III</p></div>
 <div style="background-color:#4472C4;border:3.0pt solid #000000;border-top:0;padding:1pt 4pt 1pt 4pt">
-    <p>JJJ</p>
-</div>
-<div style="background-color:#70AD47">
-    <p>KKK</p>
-    <p>LLL</p>
-</div>
-<div style="background-color:#ED7D31;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
-    <p>MMM</p>
+    <p>JJJ</p></div>
+<div style="background-color:#70AD47"><p>KKK</p>
+    <p>LLL</p></div>
+<div style="background-color:#ED7D31;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt"><p>
+    MMM</p>
     <div style="background-color:#ED7D31;border-top:3.0pt solid #000000;margin:1pt -4pt 1pt -4pt"></div>
-    <p>NNN</p>
-</div>
+    <p>NNN</p></div>
 <div style="background-color:#70AD47;border:3.0pt solid #000000;border-top:0;padding:1pt 4pt 1pt 4pt">
     <p>OOO</p>
     <div style="background-color:#70AD47;border-top:3.0pt solid #000000;margin:1pt -4pt 1pt -4pt"></div>
-    <p>PPP</p>
-</div>
+    <p>PPP</p></div>
 <div style="background-color:#00B0F0;border-bottom:3.0pt solid #000000;border-left:3.0pt solid #000000;border-right:3.0pt solid #000000;border-top:3.0pt solid #ED7D31;padding:1pt 4pt 1pt 4pt">
     <p>QQQ</p>
-    <p>RRR</p>
-</div>
+    <p>RRR</p></div>
 <div style="border-bottom:3.0pt solid #ED7D31;border-left:3.0pt solid #000000;border-right:3.0pt solid #000000;border-top:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
-    <p>SSS</p>
-</div>
+    <p>SSS</p></div>
 <table border="1">
     <tr>
         <td>
-            <div style="background-color:#A5A5A5">
-                TTT
-                <br />
-                UUU
-            </div>
+            <div style="background-color:#A5A5A5">TTT<br />UUU</div>
         </td>
         <td>
-            <div style="background-color:#ED7D31">
-                VVV
-                <br />
-                WWW
-            </div>
+            <div style="background-color:#ED7D31">VVV<br />WWW</div>
         </td>
     </tr>
     <tr>
+        <td>X<span style="background-color:#70AD47">X</span>X</td>
         <td>
-            X
-            <span style="background-color:#70AD47">X</span>
-            X
-        </td>
-        <td>
-            <div style="background-color:#2E74B5">YYY <span style="background-color:#ED7D31">ZZZ</span></div>
+            <div style="background-color:#2E74B5">YYY <span style="background-color:#ED7D31">ZZZ</span>
+            </div>
         </td>
     </tr>
 </table>
+<div style="background-color:#FFFFFF">
+    <p>Alpha</p>
+    <p>Beta</p>
+</div>
+<ol class="pydocx-list-style-type-decimal">
+    <li>
+        <div style="background-color:#70AD47;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
+            Gamma
+        </div>
+        <ol class="pydocx-list-style-type-lowerLetter">
+            <li>
+                <div style="background-color:#70AD47;border:3.0pt solid #000000;padding:1pt 4pt 1pt 4pt">
+                    Delta
+                </div>
+            </li>
+        </ol>
+    </li>
+    <li>
+        <div style="background-color:#8496B0;border:1.5pt double #FFC000;box-shadow:1.5pt 1.5pt;padding:1pt 4pt 1pt 4pt">
+            Epsilon
+        </div>
+    </li>
+</ol>
